+++ prov/efa/Makefile.include
@@ -52,7 +52,8 @@ _efa_files = \
 	prov/efa/src/rxr/rxr_pkt_entry.c \
 	prov/efa/src/rxr/rxr_pkt_type_rts.c \
 	prov/efa/src/rxr/rxr_pkt_type_data.c \
-	prov/efa/src/rxr/rxr_pkt_type_misc.c
+	prov/efa/src/rxr/rxr_pkt_type_misc.c \
+	prov/efa/src/rxr/rxr_pkt_cmd.c
 
 _efa_headers = \
 	prov/efa/src/efa.h \
@@ -61,7 +62,8 @@ _efa_headers = \
 	prov/efa/src/rxr/rxr_rma.h \
 	prov/efa/src/rxr/rxr_msg.h \
 	prov/efa/src/rxr/rxr_pkt_entry.h \
-	prov/efa/src/rxr/rxr_pkt_type.h
+	prov/efa/src/rxr/rxr_pkt_type.h \
+	prov/efa/src/rxr/rxr_pkt_cmd.h
 
 efa_CPPFLAGS += \
 	-I$(top_srcdir)/prov/efa/src/ \
