+++ prov/efa/src/rxr/rxr_cq.c
@@ -754,249 +754,6 @@ void rxr_cq_write_tx_completion(struct rxr_ep *ep,
 	return;
 }
 
-fi_addr_t rxr_cq_insert_addr_from_rts(struct rxr_ep *ep, struct rxr_pkt_entry *pkt_entry)
-{
-	int i, ret;
-	void *raw_address;
-	fi_addr_t rdm_addr;
-	struct rxr_rts_hdr *rts_hdr;
-	struct efa_ep *efa_ep;
-
-	assert(rxr_get_base_hdr(pkt_entry->pkt)->type == RXR_RTS_PKT);
-
-	efa_ep = container_of(ep->rdm_ep, struct efa_ep, util_ep.ep_fid);
-	rts_hdr = rxr_get_rts_hdr(pkt_entry->pkt);
-	assert(rts_hdr->flags & RXR_REMOTE_SRC_ADDR);
-	assert(rts_hdr->addrlen > 0);
-	if (rxr_get_base_hdr(pkt_entry->pkt)->version !=
-	    RXR_PROTOCOL_VERSION) {
-		char buffer[ep->core_addrlen * 3];
-		int length = 0;
-
-		for (i = 0; i < ep->core_addrlen; i++)
-			length += sprintf(&buffer[length], "%02x ",
-					  ep->core_addr[i]);
-		FI_WARN(&rxr_prov, FI_LOG_CQ,
-			"Host %s:Invalid protocol version %d. Expected protocol version %d.\n",
-			buffer,
-			rxr_get_base_hdr(pkt_entry->pkt)->version,
-			RXR_PROTOCOL_VERSION);
-		efa_eq_write_error(&ep->util_ep, FI_EIO, -FI_EINVAL);
-		fprintf(stderr, "Invalid protocol version %d. Expected protocol version %d. %s:%d\n",
-			rxr_get_base_hdr(pkt_entry->pkt)->version,
-			RXR_PROTOCOL_VERSION, __FILE__, __LINE__);
-		abort();
-	}
-
-	raw_address = (rts_hdr->flags & RXR_REMOTE_CQ_DATA) ?
-		      rxr_get_ctrl_cq_pkt(rts_hdr)->data
-		      : rxr_get_ctrl_pkt(rts_hdr)->data;
-
-	ret = efa_av_insert_addr(efa_ep->av, (struct efa_ep_addr *)raw_address,
-				&rdm_addr, 0, NULL);
-	if (OFI_UNLIKELY(ret != 0)) {
-		efa_eq_write_error(&ep->util_ep, FI_EINVAL, ret);
-		return -1;
-	}
-
-	return rdm_addr;
-}
-
-void rxr_cq_handle_pkt_recv_completion(struct rxr_ep *ep,
-				       struct fi_cq_data_entry *cq_entry,
-				       fi_addr_t src_addr)
-{
-	struct rxr_peer *peer;
-	struct rxr_pkt_entry *pkt_entry;
-
-	pkt_entry = (struct rxr_pkt_entry *)cq_entry->op_context;
-
-#if ENABLE_DEBUG
-	dlist_remove(&pkt_entry->dbg_entry);
-	dlist_insert_tail(&pkt_entry->dbg_entry, &ep->rx_pkt_list);
-#ifdef ENABLE_RXR_PKT_DUMP
-	rxr_ep_print_pkt("Received", ep, (struct rxr_base_hdr *)pkt_entry->pkt);
-#endif
-#endif
-	if (OFI_UNLIKELY(src_addr == FI_ADDR_NOTAVAIL))
-		pkt_entry->addr = rxr_cq_insert_addr_from_rts(ep, pkt_entry);
-	else
-		pkt_entry->addr = src_addr;
-
-	assert(rxr_get_base_hdr(pkt_entry->pkt)->version ==
-	       RXR_PROTOCOL_VERSION);
-
-	peer = rxr_ep_get_peer(ep, pkt_entry->addr);
-
-	if (rxr_env.enable_shm_transfer && peer->is_local)
-		ep->posted_bufs_shm--;
-	else
-		ep->posted_bufs_efa--;
-
-	switch (rxr_get_base_hdr(pkt_entry->pkt)->type) {
-	case RXR_RTS_PKT:
-		rxr_pkt_handle_rts_recv(ep, pkt_entry);
-		return;
-	case RXR_EOR_PKT:
-		rxr_pkt_handle_eor_recv(ep, pkt_entry);
-		return;
-	case RXR_CONNACK_PKT:
-		rxr_pkt_handle_connack_recv(ep, pkt_entry, src_addr);
-		return;
-	case RXR_CTS_PKT:
-		rxr_pkt_handle_cts_recv(ep, pkt_entry);
-		return;
-	case RXR_DATA_PKT:
-		rxr_pkt_handle_data_recv(ep, pkt_entry);
-		return;
-	case RXR_READRSP_PKT:
-		rxr_pkt_handle_readrsp_recv(ep, cq_entry, pkt_entry);
-		return;
-	default:
-		FI_WARN(&rxr_prov, FI_LOG_CQ,
-			"invalid control pkt type %d\n",
-			rxr_get_base_hdr(pkt_entry->pkt)->type);
-		assert(0 && "invalid control pkt type");
-		rxr_cq_handle_cq_error(ep, -FI_EIO);
-		return;
-	}
-	return;
-}
-
-static int rxr_send_completion_mr_dereg(struct rxr_tx_entry *tx_entry)
-{
-	int i, ret = 0;
-
-	for (i = tx_entry->iov_mr_start; i < tx_entry->iov_count; i++) {
-		if (tx_entry->mr[i]) {
-			ret = fi_close((struct fid *)tx_entry->mr[i]);
-			if (OFI_UNLIKELY(ret))
-				return ret;
-		}
-	}
-	return ret;
-}
-
-void rxr_cq_handle_pkt_send_completion(struct rxr_ep *ep, struct fi_cq_data_entry *comp)
-{
-	struct rxr_pkt_entry *pkt_entry;
-	struct rxr_tx_entry *tx_entry = NULL;
-	struct rxr_peer *peer;
-	struct rxr_rts_hdr *rts_hdr = NULL;
-	struct rxr_readrsp_hdr *readrsp_hdr = NULL;
-	uint32_t tx_id;
-	int ret;
-
-	pkt_entry = (struct rxr_pkt_entry *)comp->op_context;
-	assert(rxr_get_base_hdr(pkt_entry->pkt)->version ==
-	       RXR_PROTOCOL_VERSION);
-	peer = rxr_ep_get_peer(ep, pkt_entry->addr);
-
-	switch (rxr_get_base_hdr(pkt_entry->pkt)->type) {
-	case RXR_RTS_PKT:
-		/*
-		 * for FI_READ, it is possible (though does not happen very offen) that at the point
-		 * tx_entry has been released. The reason is, for FI_READ:
-		 *     1. only the initator side will send a RTS.
-		 *     2. the initator side will receive data packet. When all data was received,
-		 *        it will release the tx_entry
-		 * Therefore, if it so happens that all data was received before we got the send
-		 * completion notice, we will have a released tx_entry at this point.
-		 * Nonetheless, because for FI_READ tx_entry will be release in rxr_handle_rx_completion,
-		 * we will ignore it here.
-		 *
-		 * For shm provider, we will write completion for small & medium  message, as data has
-		 * been sent in the RTS packet; for large message, will wait for the EOR packet
-		 */
-		rts_hdr = rxr_get_rts_hdr(pkt_entry->pkt);
-		if (!(rts_hdr->flags & RXR_READ_REQ)) {
-			tx_id = rts_hdr->tx_id;
-			tx_entry = ofi_bufpool_get_ibuf(ep->tx_entry_pool, tx_id);
-			tx_entry->bytes_acked += rxr_get_rts_data_size(ep, rts_hdr);
-		}
-		break;
-	case RXR_CONNACK_PKT:
-		break;
-	case RXR_CTS_PKT:
-		break;
-	case RXR_DATA_PKT:
-		tx_entry = (struct rxr_tx_entry *)pkt_entry->x_entry;
-		tx_entry->bytes_acked +=
-			rxr_get_data_pkt(pkt_entry->pkt)->hdr.seg_size;
-		break;
-	case RXR_READRSP_PKT:
-		readrsp_hdr = rxr_get_readrsp_hdr(pkt_entry->pkt);
-		tx_id = readrsp_hdr->tx_id;
-		tx_entry = ofi_bufpool_get_ibuf(ep->readrsp_tx_entry_pool, tx_id);
-		assert(tx_entry->cq_entry.flags & FI_READ);
-		tx_entry->bytes_acked += readrsp_hdr->seg_size;
-		break;
-	case RXR_RMA_CONTEXT_PKT:
-		rxr_pkt_handle_rma_context_send_completion(ep, pkt_entry);
-		return;
-	default:
-		FI_WARN(&rxr_prov, FI_LOG_CQ,
-			"invalid control pkt type %d\n",
-			rxr_get_base_hdr(pkt_entry->pkt)->type);
-		assert(0 && "invalid control pkt type");
-		rxr_cq_handle_cq_error(ep, -FI_EIO);
-		return;
-	}
-
-	if (tx_entry && tx_entry->total_len == tx_entry->bytes_acked) {
-		if (tx_entry->state == RXR_TX_SEND)
-			dlist_remove(&tx_entry->entry);
-		if (tx_entry->state == RXR_TX_SEND &&
-		    efa_mr_cache_enable && rxr_ep_mr_local(ep)) {
-			ret = rxr_send_completion_mr_dereg(tx_entry);
-			if (OFI_UNLIKELY(ret)) {
-				FI_WARN(&rxr_prov, FI_LOG_MR,
-					"In-line memory deregistration failed with error: %s.\n",
-					fi_strerror(-ret));
-			}
-		}
-
-		peer->tx_credits += tx_entry->credit_allocated;
-
-		if (tx_entry->cq_entry.flags & FI_READ) {
-			/*
-			 * this must be on remote side
-			 * see explaination on rxr_cq_handle_rx_completion
-			 */
-			struct rxr_rx_entry *rx_entry = NULL;
-
-			rx_entry = ofi_bufpool_get_ibuf(ep->rx_entry_pool, tx_entry->rma_loc_rx_id);
-			assert(rx_entry);
-			assert(rx_entry->state == RXR_RX_WAIT_READ_FINISH);
-
-			if (ep->util_ep.caps & FI_RMA_EVENT) {
-				rx_entry->cq_entry.len = rx_entry->total_len;
-				rx_entry->bytes_done = rx_entry->total_len;
-				efa_cntr_report_rx_completion(&ep->util_ep, rx_entry->cq_entry.flags);
-			}
-
-			rxr_release_rx_entry(ep, rx_entry);
-			/* just release tx, do not write completion */
-			rxr_release_tx_entry(ep, tx_entry);
-		} else if (tx_entry->cq_entry.flags & FI_WRITE) {
-			if (tx_entry->fi_flags & FI_COMPLETION) {
-				rxr_cq_write_tx_completion(ep, tx_entry);
-			} else {
-				efa_cntr_report_tx_completion(&ep->util_ep, tx_entry->cq_entry.flags);
-				rxr_release_tx_entry(ep, tx_entry);
-			}
-		} else {
-			assert(tx_entry->cq_entry.flags & FI_SEND);
-			rxr_cq_write_tx_completion(ep, tx_entry);
-		}
-	}
-
-	rxr_pkt_entry_release_tx(ep, pkt_entry);
-	if (!peer->is_local)
-		rxr_ep_dec_tx_pending(ep, peer, 0);
-	return;
-}
-
 static int rxr_cq_close(struct fid *fid)
 {
 	int ret;
